change run configuration and load env procedure
